added basic deserialization capabily to command code gen
