+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/java/client/SpringClientRenderer.kt
@@ -4,7 +4,7 @@ import com.google.inject.Inject
 import io.vrap.codegen.languages.java.JavaSubTemplates
 import io.vrap.codegen.languages.java.extensions.EObjectTypeExtensions
 import io.vrap.codegen.languages.java.extensions.fullClassName
-import io.vrap.codegen.languages.java.extensions.simpleName
+import io.vrap.codegen.languages.java.extensions.returnType
 import io.vrap.codegen.languages.java.extensions.toComment
 import io.vrap.rmf.codegen.common.generator.core.ResourceCollection
 import io.vrap.rmf.codegen.io.TemplateFile
@@ -17,9 +17,6 @@ import io.vrap.rmf.codegen.types.VrapObjectType
 import io.vrap.rmf.codegen.types.VrapTypeProvider
 import io.vrap.rmf.raml.model.resources.Method
 import io.vrap.rmf.raml.model.resources.Resource
-import io.vrap.rmf.raml.model.responses.Response
-import io.vrap.rmf.raml.model.types.AnyType
-import io.vrap.rmf.raml.model.types.impl.TypesFactoryImpl
 
 class SpringClientRenderer @Inject constructor(val packageProvider: PackageProvider, override val vrapTypeProvider: VrapTypeProvider) : ResourceCollectionRenderer, EObjectTypeExtensions {
 
@@ -79,7 +76,7 @@ class SpringClientRenderer @Inject constructor(val packageProvider: PackageProvi
 
     fun javaBody(resource: Resource, method: Method): String {
 
-        val methodReturnType = vrapTypeProvider.doSwitch(method.retyurnType())
+        val methodReturnType = vrapTypeProvider.doSwitch(method.returnType())
         val returnIsEmpty = methodReturnType is VrapNilType
 
         val body = """
@@ -97,7 +94,7 @@ class SpringClientRenderer @Inject constructor(val packageProvider: PackageProvi
             |
             |    <${method.mediaType().escapeAll()}>
             |
-            |    ${if(returnIsEmpty) "final ParameterizedTypeReference\\<Void\\> type = new ParameterizedTypeReference\\<Void\\>() {};" else "final ParameterizedTypeReference\\<${method.retyurnType().toVrapType().fullClassName().escapeAll()}\\> type = new ParameterizedTypeReference\\<${method.retyurnType().toVrapType().fullClassName().escapeAll()}\\>() {};"}
+            |    ${if(returnIsEmpty) "final ParameterizedTypeReference\\<Void\\> type = new ParameterizedTypeReference\\<Void\\>() {};" else "final ParameterizedTypeReference\\<${method.returnType().toVrapType().fullClassName().escapeAll()}\\> type = new ParameterizedTypeReference\\<${method.returnType().toVrapType().fullClassName().escapeAll()}\\>() {};"}
             |    final String fullUri = baseUri + "${relativeUrl(resource,method)}";
             |
             |    ${if(returnIsEmpty) "return restTemplate.exchange(fullUri, HttpMethod.${method.method.name.toUpperCase()}, entity, type, parameters);" else "return restTemplate.exchange(fullUri, HttpMethod.${method.method.name.toUpperCase()}, entity, type, parameters).getBody();"}
@@ -157,17 +154,4 @@ class SpringClientRenderer @Inject constructor(val packageProvider: PackageProvi
         return "final HttpEntity<?> entity = null;"
 
     }
-
-    fun Method.retyurnType(): AnyType {
-        return this.responses
-                .filter { it.isSuccessfull() }
-                .filter { it.bodies?.isNotEmpty() ?: false }
-                .firstOrNull()
-                ?.let { it.bodies[0].type }
-                ?: TypesFactoryImpl.eINSTANCE.createNilType()
-    }
-
-
-    fun Response.isSuccessfull(): Boolean = this.statusCode.toInt() in (200..299)
-
 }
