+++ codegen-renderers/src/main/kotlin/io/vrap/codegen/languages/java/extensions/MethodExtensions.kt
@@ -4,6 +4,9 @@ import com.damnhandy.uri.template.Expression
 import com.damnhandy.uri.template.UriTemplate
 import io.vrap.rmf.raml.model.resources.Method
 import io.vrap.rmf.raml.model.resources.Resource
+import io.vrap.rmf.raml.model.responses.Response
+import io.vrap.rmf.raml.model.types.AnyType
+import io.vrap.rmf.raml.model.types.impl.TypesFactoryImpl
 import io.vrap.rmf.raml.model.util.StringCaseFormat
 import java.util.stream.Collectors
 
@@ -12,6 +15,17 @@ fun Method.toRequestName(): String {
 }
 
 fun Method.resource(): Resource = this.eContainer() as Resource
+
+fun Method.returnType(): AnyType {
+    return this.responses
+            .filter { it.isSuccessfull() }
+            .filter { it.bodies?.isNotEmpty() ?: false }
+            .firstOrNull()
+            ?.let { it.bodies[0].type }
+            ?: TypesFactoryImpl.eINSTANCE.createNilType()
+}
+
+fun Response.isSuccessfull(): Boolean = this.statusCode.toInt() in (200..299)
 
 fun UriTemplate.toParamName(delimiter: String): String {
     return this.toParamName(delimiter, "")
